Fix two bugs
Fixed two bugs in the Brooks2015 analysis code that don't affect any
reported results. First, the predicted proportions in the data frame
`main0` were being incorrectly calculated to be the calibratd
proportions (but were correctly calculated in `pred`, so the results are
unaffected). Second, is that the reduction in Aitchison error by the
fitted model was being incorrectly calculated (though again doesn't
affect the reported results).

Fixes #2.